[backport] BUG: Workaround for Slicer startup hang on Windows11
numpy has to be loaded early to avoid hang when loading openblas library.
This is a temporary workaround and may be removed if the related issue is fixed in Windows11 or numpy (or we find a nicer solution in Slicer).

Cherry-picked commit https://github.com/Slicer/Slicer/commit/542f53c5ba9bf650119707f990e36d74a6c8d36a