[backport] ENH: Add ability to quickly open to log file location
This commit updates the report a bug dialog to have a button that opens the log
location instead of copying the log contents. This was done because copying log
contents is a simple thing to do (ctrl +a, ctrl + c) where as opening to log
location is far more troublesome. Going to the log location allows for copying
the entire log file for sharing the entire file as-is with others.

Cherry-picked commit https://github.com/Slicer/Slicer/commit/2395c07917e1d612ebe386e50df460cd61df52c4 from main Slicer repository.